pyproject.toml
[tool.poetry]
name = "anekrin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
asyncpg = "^0.25.0"
tortoise-orm = "^0.18.1"
matplotlib = "^3.5.1"
aerich = "^0.6.2"
aio-pika = "^7.1.0"
aiogram = "^2.19"
emoji = "^1.6.3"
sentry-sdk = "^1.5.6"
numpy = "^1.22.3"
pandas = "^1.4.1"
calmap = "^0.0.9"
pytest = "^7.1.1"
pytest-asyncio = "^0.18.3"
mplcyberpunk = "^0.5.0"

[tool.poetry.dev-dependencies]

[tool.aerich]
tortoise_orm = "app.models.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
